dmrfit: Optimization Tools for Discrete Markov Random Fields

A set of tools designed for estimating discrete Markov Random Fields via pseudolikelihood, for both Ising models (Ising, 1925, doi:10.1007/BF02980577 ) and Ordinal Markov Random Fields (Marsman, M., van den Bergh, D., and Haslbeck, J. M. B, 2025, doi:10.1017/psy.2024.4 ). Parameters are estimated through a trust region optimization algorithm, with robust standard error estimation. The package supports both full network estimation, where all edges are included, and constrained network estimation, where only a specified subset of edges is considered.

See also

Author

Maintainer: Giuseppe Arena g.arena@uva.nl (ORCID)