dmrfit

Fit a discrete Markov Random Field model via pseudo-likelihood estimation. The optimization is performed using the trust region algorithm implemented by

dmrfit(
  data,
  parinit = NULL,
  structure = NULL,
  with_prior = FALSE,
  savage_dickey = FALSE,
  M = 1000,
  oversampling = 10,
  ncores = 1,
  thresholds_alpha = 0.5,
  thresholds_beta = 0.5,
  interactions_location = 0,
  interactions_scale = 2.5,
  proposal_df = 5,
  seed = 123
)

Arguments

data

data matrix, with rows as samples and columns as variables. Each variable should be rescaled to the range of 0 to m-1, where m is the number of categories for that variable. The baseline category is always the minimum value in the variable. The internal processing will check if the variables are rescaled and will rescale them if necessary. If there are any NAs in the data, they will be removed before optimization (listwise deletion).

parinit

parinit, initial parameter estimates, a vector of length equal to the number of parameters in the model. If NULL, it will be initialized to a vector of zeros. The number of parameters in the model is calculated as sum(P * (n_categories - 1)) + P * (P - 1) / 2, where P is the number of variables.

structure

network structure, P x P matrix, with 0 for no edge and 1 for edge. If NULL, then fully connected graph is assumed. Default is NULL

with_prior

with_prior, logical, whether to include the prior in the optimization. If TRUE, a Beta-Prime and a Cauchy prior are applied to thresholds and pairwise associations respectively. If FALSE, no prior is applied. Default is FALSE

savage_dickey

logical, whether to compute the Savage-Dickey density ratio Bayes factor for each pairwise interaction via Bayesian Sampling Importance Resampling (BSIR). Only available when with_prior = TRUE. Default is FALSE.

M

number of importance samples for the BSIR step (default is 1000). Only used when savage_dickey = TRUE.

oversampling

multiplier for the number of proposal draws in the BSIR step (default is 10). The total number of proposal samples is M * oversampling. Only used when savage_dickey = TRUE.

ncores

ncores, number of cores to use for parallel computing of the gradient, hessian and likelihood values. If ncores is larger than the number of available cores, it will be set to the number of available cores minus one. Default is 1.

thresholds_alpha

alpha parameter for the Beta-Prime prior on thresholds (default is 0.5). Only used when with_prior = TRUE.

thresholds_beta

beta parameter for the Beta-Prime prior on thresholds (default is 0.5). Only used when with_prior = TRUE.

interactions_location

location parameter for the Cauchy prior on pairwise interactions (default is 0.0). Only used when with_prior = TRUE.

interactions_scale

scale parameter for the Cauchy prior on pairwise interactions (default is 2.5). Only used when with_prior = TRUE.

proposal_df

degrees of freedom for the multivariate t proposal distribution in the BSIR step (default is 5). Only used when savage_dickey = TRUE.

seed

random seed for reproducibility of the BSIR step (default is 123). Only used when savage_dickey = TRUE.

Value

dmrfit S3 class object

Examples

# simulate data from a 3-node Ising model with 2 categories per node
set.seed(123)
n <- 1000
P <- 3
structure <- matrix(c(0, 1, 0, 1, 0, 1, 0, 1, 0), nrow = P)
data <- matrix(rbinom(n * P, size = 1, prob = c(0.2, 0.5, 0.3)[rep(1:P, each = n)]), nrow = n, ncol = P)
fit <- dmrfit(data, structure = structure, with_prior = TRUE, savage_dickey = TRUE, M = 10000, ncores = 2)
summary(fit)
#> 
#> Call:
#> dmrfit(data = data, structure = structure, with_prior = TRUE, 
#>     savage_dickey = TRUE, M = 10000, ncores = 2)
#> 
#> Discrete Markov Random Field (Ising)
#> Estimation method: maximum pseudo-likelihood (with prior)
#> Network structure: constrained
#> Nodes: 3  Observations: 1000  Free parameters: 5 
#> Standard errors: Huber-White sandwich estimator
#> ------------------------------------------------------------ 
#> 
#> Thresholds:
#>         Post.Mode   Post.SD  z value  Pr(>|z|)
#> mu[1,1] -1.329430  0.117546 -11.3099 < 2.2e-16
#> mu[2,1]  0.017534  0.082013   0.2138    0.8307
#> mu[3,1] -0.809384  0.101373  -7.9842 1.414e-15
#> 
#> Pairwise interactions:
#>            Post.Mode   Post.SD z value Pr(>|z|)
#> sigma[2,1] -0.137283  0.158022 -0.8688   0.3850
#> sigma[3,2] -0.007817  0.136502 -0.0573   0.9543
#> 
#> Negative pseudo-loglikelihood: 1818.673 
#> 
#> Savage-Dickey density ratio  [prior: Cauchy( 0 , 2.5 )]
#> H0: sigma = 0 for each pairwise interaction
#> 
#>            Estimate     SE   BF_01 Pr(=0|x)
#> sigma[2,1]  -0.1373 0.1580 13.4723   0.9309
#> sigma[3,2]  -0.0078 0.1365 31.7240   0.9694
#> 
#> SIR samples: 10000   Effective sample size: 7003.1