dmrfit_bayes

Fit a discrete Markov Random Field model via pseudo-likelihood estimation. The optimization is performed using the Coordinate Rescaling sampler implemented by Arena and Marsman (2026).

dmrfit_bayes(
  data,
  parinit = NULL,
  nsim = 1000,
  burnin = 1000,
  ncores = 1,
  thresholds_alpha = 0.5,
  thresholds_beta = 0.5,
  interactions_location = 0,
  interactions_scale = 2.5,
  sigma2 = 0.1,
  seed = 123
)

Arguments

data

data matrix, with rows as samples and columns as variables. Each variable should be rescaled to the range of 0 to m-1, where m is the number of categories for that variable. The baseline category is always the minimum value in the variable. The internal processing will check if the variables are rescaled and will rescale them if necessary. If there are any NAs in the data, they will be removed before optimization (listwise deletion).

parinit

parinit, initial parameter estimates, a vector of length equal to the number of parameters in the model. If NULL, it will be initialized to a vector of zeros. The number of parameters in the model is calculated as sum(P * (n_categories - 1)) + P * (P - 1) / 2, where P is the number of variables.

nsim

nsim, number of iterations for the optimization. Default is 1000.

burnin

burnin, number of burnin iterations for the optimization. Default is 1000.

ncores

ncores, number of cores to use for parallel computing of the gradient, hessian and likelihood values. If ncores is larger than the number of available cores, it will be set to the number of available cores minus one. Default is 1.

thresholds_alpha

alpha parameter for the Beta-Prime prior on thresholds (default is 0.5).

thresholds_beta

beta parameter for the Beta-Prime prior on thresholds (default is 0.5).

interactions_location

location parameter for the Cauchy prior on pairwise interactions (default is 0.0).

interactions_scale

scale parameter for the Cauchy prior on pairwise interactions (default is 2.5).

sigma2

initial value for the adaptive variance parameter in the FisherMALA sampler (default is 0.1).

seed

random seed for reproducibility of the BSIR step (default is 123).

Value

dmrfit S3 class object

Examples

# simulate data from a 3-node Ising model with 2 categories per node
set.seed(123)
n <- 1000
P <- 3
data <- matrix(rbinom(n * P, size = 1, prob = c(0.2, 0.5, 0.3)[rep(1:P, each = n)]), nrow = n, ncol = P)
fit <- dmrfit_bayes(data = data, nsim = 500, burnin = 500)
summary(fit)
#> 
#> Call:
#> dmrfit_bayes(data = data, nsim = 500, burnin = 500)
#> 
#> Discrete Markov Random Field (Ising)
#> Estimation method: Bayesian (FisherMALA)
#> Nodes: 3  Observations: 1000  Parameters: 6 
#> MCMC samples: 500  Acceptance rate: 0.578  Elapsed: 0.4 sec
#> ------------------------------------------------------------ 
#> 
#> Thresholds:
#>                 Post.Mean Post.SD    2.5%   97.5%
#> mu[1,1] -1.4423   -1.4613  0.1285 -1.7063 -1.2356
#> mu[2,1]  0.0164    0.0108  0.0787 -0.1473  0.1651
#> mu[3,1] -0.8828   -0.8881  0.1055 -1.1078 -0.6999
#> 
#> Pairwise interactions:
#>                    Post.Mean Post.SD    2.5%  97.5%
#> sigma[2,1] -0.1374   -0.1105  0.1765 -0.4601 0.2239
#> sigma[3,1]  0.3445    0.3635  0.1683  0.0295 0.6857
#> sigma[3,2] -0.0040   -0.0093  0.1380 -0.2818 0.3017
#> 
#> Negative pseudo-loglikelihood at posterior mode: 1814.438 
#> 
#> Savage-Dickey density ratio  [prior: Cauchy( 0 , 2.5 )]
#> H0: sigma = 0 for each pairwise interaction
#> 
#>            Post.Mode Post.SD   BF_01 Pr(=0|x)
#> sigma[2,1]   -0.1374  0.1765 14.5557   0.9357
#> sigma[3,1]    0.3445  0.1683  1.7224   0.6327
#> sigma[3,2]   -0.0040  0.1380 23.8338   0.9597
#> 
#> MCMC samples: 500   Effective sample size: 505.6